Move rangeInclusive to array-helpers utility
The rangeInclusive is clearly a utility.  It might not get used
anywhere else, but it increases the readability of the component file
to move this helper to a separate place.
