+++ app/components/page-list.js
@@ -1,23 +1,8 @@
 import { computed } from '@ember/object';
 import Component from '@ember/component';
+import { rangeInclusive } from 'ipfs-search/utils/array-helpers';
 
 const PAGES_AROUND_CURRENT = 3;
-
-/**
- * Returns an array from start to end, both inclusive.
- * If end is not greater than start, resultList is
- * returned.
- */
-function rangeInclusive( start, end ){
-  const resultList = [];
-  if( ! end > start )
-    return resultList;
-
-  for (let i = start; i <= end; i++) {
-    resultList.push(i);
-  }
-  return resultList;
-}
 
 export default Component.extend({
   pagesBeforeMe: computed('totalPages', 'currentPage', function() {
