+++ tests/unit/utils/array-helpers-test.js
@@ -1,0 +1,10 @@
+import arrayHelpers from 'ipfs-search/utils/array-helpers';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | array helpers');
+
+// Replace this with your real tests.
+test('it works', function(assert) {
+  let result = arrayHelpers();
+  assert.ok(result);
+});
